+++ lib/legion/extensions/microsoft_teams/helpers/client.rb
@@ -11,8 +11,7 @@ module Legion
                                                       Legion::Extensions::Helpers.const_defined?(:Lex, false)
 
           def graph_connection(token: nil, api_url: 'https://graph.microsoft.com/v1.0', **_opts)
-            token ||= settings&.dig(:auth, :delegated, :token)
-            token ||= TokenCache.instance.cached_delegated_token if defined?(TokenCache)
+            token ||= entra_delegated_token
             Faraday.new(url: api_url) do |conn|
               conn.request :json
               conn.response :json, content_type: /\bjson$/
@@ -22,8 +21,6 @@ module Legion
           end
 
           def bot_connection(token: nil, service_url: 'https://smba.trafficmanager.net/teams/', **_opts)
-            token ||= settings&.dig(:auth, :bot, :token)
-            token ||= TokenCache.instance.cached_app_token if defined?(TokenCache)
             Faraday.new(url: service_url) do |conn|
               conn.request :json
               conn.response :json, content_type: /\bjson$/
@@ -41,6 +38,15 @@ module Legion
               conn.request :url_encoded
               conn.response :json, content_type: /\bjson$/
             end
+          end
+
+          private
+
+          def entra_delegated_token
+            Legion::Extensions::Identity::Entra::Helpers::TokenManager.load_token(:delegated)
+          rescue StandardError => e
+            log.debug("entra_delegated_token: #{e.message}")
+            nil
           end
         end
       end