+++ spec/legion/extensions/microsoft_teams/actors/incremental_sync_spec.rb
@@ -51,6 +51,7 @@ RSpec.describe Legion::Extensions::MicrosoftTeams::Actor::IncrementalSync do
 
     it 'calls incremental_sync on runner_class when token is present' do
       allow(actor).to receive(:resolve_token).and_return('test-token')
+      allow(Legion::Settings).to receive(:[]).with(:microsoft_teams).and_return({})
       runner = Legion::Extensions::MicrosoftTeams::Runners::ProfileIngest
       expect(runner).to receive(:incremental_sync).with(
         token: 'test-token', top_people: 10, message_depth: 50